frontend-feat-3712 added driver image
