+++ Frontend/ui-customer/src/Components/DriverInfoCard/Controller.purs
@@ -93,4 +93,5 @@ type DriverInfoCardData =
   , lastMessage :: ChatView.ChatComponent
   , config :: AppConfig
   , vehicleVariant :: String
+  , driverProfileImage :: Maybe String
   }
