+++ Frontend/ui-customer/src/Screens/RideBookingFlow/HomeScreen/ComponentConfig.purs
@@ -802,6 +802,7 @@ driverInfoTransformer state =
     , lastMessage : state.data.lastMessage
     , config : state.data.config
     , vehicleVariant : cardState.vehicleVariant
+    , driverProfileImage : Nothing
     }
 
 emergencyHelpModelViewState :: ST.HomeScreenState -> EmergencyHelp.EmergencyHelpModelState
